Update patch set 1
Patch Set 1:

This change from 10.1 is kind of surprising and at once has a few interesting implications. It assumes you have fully functional secure MM heap, which d2 does not, as it refused to boot when we implemented the build.prop enforcing the security. It further assumes that if insecure MM is requested, the requesting app is retarded, not because it's something you need. It then proceeds to hand the app an IOMMU heap instead. Again, it assumes you have a 100% fully functional IOMMU, which (again) apparently d2 does not.

Secure MM is intended for DRM-enabled video content, and enforcing it is Google's way of enticing providers like Netflix to provide HD to Android. Nevermind that nobody in their right mind is going to believe a streaming video rip of any kind is a true substitute for a real and easier to obtain HD rip, such as a copy of a blueray.

An alternative solution if allowing insecure MM heaps is not where we want to go, is to fix the fallback logic and assign the insecure heap from somewhere besides IOMMU, such as a camera heap, which would have to be added to the d2 board file.

Patch-set: 1