Update patch set 3
Patch Set 3:

Wow, funny story here.

I reverted the two display patches and the NEW_ION CFLAG on the 8660 device I'm working on, recompiled, no boot. Switched it back and still no boot! Ended up having to write all zeros to the boot partition and rewrite the image, I guess this device is dying :(

Anyway. I'm currently working on msm8660 using a 3.0 kernel, and with the NEW_ION patches reverted and the flag removed, it does work flawlessly.

I suspect you have an issue in your kernel, or in your compat code.

Assuming you are up-to-date with the latest 3.0 from CAF (jb_1.9.6 is a good start), including their mini-ion upgrade, you should just need one patch. See here for one that I know works:

https://github.com/CyanogenMod/android_kernel_samsung_msm8660-common/commits/next/drivers/gpu/ion

If you have a dmesg/logcat, I can have a look.

That said, the code that's currently in this driver is wrong! I don't see anything wrong with having an "old" behavior, but what we have in display-caf is totally wrong.

Patch-set: 3